+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## Release 0.1.0
 
 **Features**
+- Added: Able to install PowerShell 7
 
 **Bugfixes**